restructuring of code into files
