+++ lib/pages/favorites_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:nestify/widgets/PreviewCard.dart';
 import 'package:provider/provider.dart';
-import '../providers/model.dart';
+import '../models/model.dart';
 
 class FavoritesPage extends StatelessWidget {
   const FavoritesPage({super.key});
