+++ lib/widgets/PreviewCard.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:nestify/models/blueprint_post.dart';
 import 'package:go_router/go_router.dart';
 import 'package:nestify/widgets/favorite_icon_button.dart';
-import 'package:nestify/providers/model.dart';
+import 'package:nestify/models/model.dart';
 import 'package:nestify/widgets/options_menu.dart';
 import 'package:provider/provider.dart';
