+++ lib/widgets/favorite_icon_button.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:nestify/models/blueprint_post.dart';
-import 'package:nestify/providers/model.dart';
+import 'package:nestify/models/model.dart';
 import 'package:provider/provider.dart';
 
 class FavoriteIconButton extends StatefulWidget {
